+++ package.json
@@ -19,6 +19,9 @@
   "homepage": "https://github.com/Juakata/Battleship-JS#readme",
   "devDependencies": {
     "css-loader": "^3.4.2",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-plugin-import": "^2.20.0",
     "file-loader": "^5.0.2",
     "style-loader": "^1.1.2",
     "webpack": "^4.41.5",
