+++ src/style.css
@@ -83,11 +83,6 @@ header {
   margin: 0 auto;
   position: relative;
 }
-.player-ship-list h {
-}
-
-.computer-ship-list {
-}
 
 .radar {
   top: 6%;
